working staging

+++ src/Uhuru.CloudFoundry.DEA/StagingTask.cs
@@ -12,6 +12,7 @@ namespace Uhuru.CloudFoundry.DEA
     using System.IO;
     using System.Linq;
     using System.Net;
+    using System.Security.Cryptography;
     using System.Text;
     using Uhuru.CloudFoundry.DEA.Messages;
     using Uhuru.Utilities;
@@ -27,7 +28,7 @@ namespace Uhuru.CloudFoundry.DEA
         public string TaskId { get; set; }
         public string StreamingLogUrl { get; set; }
         public string TaskLog { get; set; }
-        public string DetectedBuildpack { get; set; }
+        public string DetectedBuildpack { get { return StagingInfo.GetDetectedBuildpack(Path.Combine(this.workspace.StagedDir, StagingWorkspace.StagingInfo)); } }
         public string DropletSHA { get; set; }
 
         private StagingStartMessageRequest message;
@@ -38,14 +39,29 @@ namespace Uhuru.CloudFoundry.DEA
         {
             this.TaskId = message.TaskID;
             this.message = message;
-            this.workspace = new StagingWorkspace(dropletDir);
+            this.workspace = new StagingWorkspace(dropletDir, message.TaskID);
             this.buildpacksDir = buildpacksDirectory;
         }
 
         public void Start() 
         {
-            StagingSetup();
-            Staging();
+            try
+            {
+                StagingSetup();
+                Staging();
+                UploadDroplet();
+                SaveBuildpackCache();
+                AfterUpload(null);
+            }
+            catch(Exception ex)
+            {
+                AfterUpload(ex);
+                throw ex;
+            }
+            finally
+            {                
+                Directory.Delete(this.workspace.WorkspaceDir, true);
+            }
         }
 
         private void StagingSetup()
@@ -57,33 +73,39 @@ namespace Uhuru.CloudFoundry.DEA
                 {
                     DownloadBuildpackCache();
                 }
-                File.Create(this.workspace.StagingLogPath);                
+                Directory.CreateDirectory(new FileInfo(this.workspace.StagingLogPath).DirectoryName);
+                if(!File.Exists(this.workspace.StagingLogPath))
+                    File.Create(this.workspace.StagingLogPath);
+                AfterSetup(null);
             }
             catch (Exception ex)
             {
                 AfterSetup(ex);
-            }
-            finally
-            {
-                AfterSetup(null);
-            }
+                throw ex;
+            }            
         }
 
         private void Staging()
-        {
+        {            
             UnpackApp();
             UnpackBuildpackCache();
             Stage();
             PackApp();
+            
+            Directory.CreateDirectory(this.workspace.StagedDropletDir);
+            File.Copy(this.workspace.StagedDroplet, this.workspace.StagedDropletPath);
 
-            File.Copy(this.workspace.StagedDroplet, Path.Combine(this.workspace.StagedDropletDir, Path.GetFileName(this.workspace.StagedDroplet)));
-
-            LogUploadStarted();
-            StagingInfo();
+            using (Stream stream = File.OpenRead(this.workspace.StagedDropletPath))
+            {
+                using (SHA1 sha = SHA1.Create())
+                {
+                    this.DropletSHA = BitConverter.ToString(sha.ComputeHash(stream)).Replace("-", string.Empty);
+                }
+            }
         }
 
         private void DownloadApp()
-        {
+        {            
             WebClient client = new WebClient();
             try
             {
@@ -124,15 +146,16 @@ namespace Uhuru.CloudFoundry.DEA
 
         private void UnpackApp() 
         {
+            Directory.CreateDirectory(this.workspace.UnstagedDir);
             DEAUtilities.UnzipFile(this.workspace.UnstagedDir, this.workspace.DownloadDropletPath);
         }
         private void UnpackBuildpackCache() 
         {
+            Directory.CreateDirectory(this.workspace.Cache);
             if (File.Exists(this.workspace.DownloadBuildpackCachePath))
             {
-                string tarFileName = Path.GetFileName(this.workspace.DownloadBuildpackCachePath);
-                tarFileName = Path.ChangeExtension(tarFileName, ".tar");
                 DEAUtilities.UnzipFile(this.workspace.Cache, this.workspace.DownloadBuildpackCachePath); // Unzip
+                string tarFileName = Directory.GetFiles(this.workspace.DownloadBuildpackCachePath, "*.tar")[0];
                 DEAUtilities.UnzipFile(this.workspace.Cache, Path.Combine(this.workspace.Cache, tarFileName)); // Untar
                 File.Delete(Path.Combine(this.workspace.Cache, tarFileName));
             }
@@ -149,8 +172,9 @@ namespace Uhuru.CloudFoundry.DEA
             Buildpack buildpack = null;
             if (this.message.Properties.Buildpack != null)
             {
+                Directory.CreateDirectory(Path.Combine(this.workspace.TempDir, "buildpacks"));
                 string buildpackPath = Path.Combine(this.workspace.TempDir, "buildpacks", Path.GetFileName(new Uri(this.message.Properties.Buildpack).LocalPath));
-                string command = string.Format("git clone --recursive {0} {1}", this.message.Properties.Buildpack, buildpackPath);
+                string command = string.Format("\"E:\\Program Files (x86)\\Git\\bin\\git.exe\" clone --recursive {0} {1}", this.message.Properties.Buildpack, buildpackPath);
                 int success = DEAUtilities.ExecuteCommand(command);
                 if (success != 0)
                 {
@@ -177,7 +201,7 @@ namespace Uhuru.CloudFoundry.DEA
             }
             buildpack.Compile(); // TODO need to add timeout
 
-            
+            StagingInfo.SaveBuildpackInfo(Path.Combine(this.workspace.StagedDir, StagingWorkspace.StagingInfo), buildpack.Name, GetStartCommand(buildpack));
 
         }
 
@@ -189,14 +213,55 @@ namespace Uhuru.CloudFoundry.DEA
             File.Delete(tempFile);
         }
 
-        private void CopyOut() 
+        private string GetStartCommand(Buildpack buildpack)
         {
-            
+            if (this.message.Properties.Meta != null)
+            {
+                if (this.message.Properties.Meta.Command != null)
+                {
+                    return this.message.Properties.Meta.Command;
+                }
+            }
+            ReleaseInfo info = buildpack.GetReleaseInfo();
+            if (info.defaultProcessType != null)
+            {
+                if (info.defaultProcessType.Web != null)
+                {
+                    return info.defaultProcessType.Web;
+                }
+            }
+            throw new Exception("Please specify a web start command in your manifest.yml");
         }
-        private void SaveDroplet() { }
-        private void LogUploadStarted() { }
-        private void StagingInfo() { }
 
+        private void UploadDroplet() 
+        {
+            Uri uri = new Uri(this.message.UploadURI);
+            DEAUtilities.HttpUploadFile(this.message.UploadURI, new FileInfo(this.workspace.StagedDropletPath), "upload[droplet]", "application/octet-stream", uri.UserInfo);
+        }
+
+        private void SaveBuildpackCache()
+        {
+            try
+            {
+                PackBuildpackCache();
+            }
+            catch
+            {
+                return;
+            }
+            File.Copy(this.workspace.StagedBuildpackCache, this.workspace.StagedBuildpackCachePath);
+            Uri uri = new Uri(this.message.BuildpackCacheUploadURI);
+            DEAUtilities.HttpUploadFile(this.message.BuildpackCacheUploadURI, new FileInfo(this.workspace.StagedBuildpackCachePath), "upload[droplet]", "application/octet-stream", uri.UserInfo); 
+        }
+
+        private void PackBuildpackCache()
+        {
+            Directory.CreateDirectory(this.workspace.Cache);
+            string tempFile = Path.ChangeExtension(this.workspace.StagedBuildpackCache, "tar");
+            DEAUtilities.TarDirectory(this.workspace.Cache, tempFile);
+            DEAUtilities.GzipFile(tempFile, this.workspace.StagedBuildpackCache);
+            File.Delete(tempFile);
+        }
         
     }
 }
